sdk/rust: shared accountant ModificationKind
