+++ cosmwasm/contracts/ntt-global-accountant/src/contract.rs
@@ -28,6 +28,7 @@ use tinyvec::{Array, TinyVec};
 use wormhole_bindings::WormholeQuery;
 use wormhole_io::TypePrefixedPayload;
 use wormhole_sdk::{
+    accountant_modification::ModificationKind,
     ntt_accountant as ntt_accountant_module, relayer,
     vaa::{self, Body, Header, Signature},
     Chain,
@@ -465,9 +466,9 @@ fn handle_accountant_governance_vaa(
         } => {
             let token_address = TokenAddress::new(token_address.0);
             let kind = match kind {
-                ntt_accountant_module::ModificationKind::Add => Kind::Add,
-                ntt_accountant_module::ModificationKind::Subtract => Kind::Sub,
-                ntt_accountant_module::ModificationKind::Unknown => {
+                ModificationKind::Add => Kind::Add,
+                ModificationKind::Subtract => Kind::Sub,
+                ModificationKind::Unknown => {
                     bail!("unsupported governance action")
                 }
             };
